requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
lxml==4.9.3
python-multipart==0.0.6
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.35.2
torch==2.1.1


